doc: Add Design Considerations doc, move implementation info and examples there
The usage examples and implementations didn't really felt like they
fit in well with the style of the protocol documentation. It may be
better to just present the protocol, then after people have seen it,
give them an idea of how it could (but not necessarily should) be
used. Added a Design Considerations doc for this work.