Updating the version numbers
We will no longer try and represent the schema or the protocol
documentation with SimVer like release numbers, instead we will
use the simple unsigned int schema message version.

Updates to the docs may happen without increaing this number, as
the docs here are driven by the schema, which cannot change without
incrementing the message version.