feat: Make clients and devices work with clone trait
When a client is created, it should create a new future along with
itself. The future will represent the event loop, and the client will
access the event loop. This means that clients and devices can be
cloned as needed, while still ensuring they can talk to the original
event loop.