chore: Make client event loop handle devices
In C# and JS, we usually let devices have a reference to the client
instance, which they used to communicate to the server. Since handing
client references to every devices in Rust is a lot more work, we let
the client event loop handle device management and communication
directly. Devices can be queries from the client, which will request
them from the inner loop.

The reason for this is that devices can be dropped by the application,
but still connected to the server. We want applications to be able to
get active devices at any time, but we can't easily just hand out a
ton of references like we do in GC'd languages, since we don't really
have EventHandlers to spread events and it'd required Arc::Mutex'ing a
bunch of the internals anyways.

This is complicated, but overall I think it works.