chore: Make client/device check events on all event loop comms
Since we're using async_std channels, if we send a message and the
event loop has shut down, we may never know if we don't check the
status of an event loop receiver to see if it's returned None. Always
run checks before sending messages to the event loop.

Fixes #25