feat: Add beginnings of protocols, overhaul Rumble handling
This patch is way, way too big, but it snowballed so it's just gonna
land this way.

Implements the base traits and a basic Lovense protocol that just
handles vibration. This is mostly to just get something up and running
to make sure the system flow is correct.

While we're at it, overhaul the Rumble BLE comm manager to actually
deal with futures like its supposed to, versus setting up channels
everywhere. As long as we're expecting replies (which we will be from
outside commands), we should just be able to pipe in futures and have
the rumble thread resolve them. We only need a channel outward in the
cases where we expect notifications.