feat: Adding Kiiroo v2 and v2.1 protocols
The Onyx 2.1 and Titan 2.1 are untested because I own neither.

LinearCmd is currently failing hard: the way that DeviceManager
handles device messages is by cloning the device object. This
means every time a protocol impl function is called, it's being
called on a copy of the original object: no state is recorded.

This is an issue for all devices, but so far we've only seen it
with LinearCmd because we've yet to actually test that duplicate
instructions are dropped, etc.