chore: Don't require waker passing in public client traits
The division between futures and "State" types in Buttplug is hard to
understand and should stay internal as we pass things across
channels/thread/etc... This patch removes the requirement of passing
state from the ClientConnector trait's send method, treating it as a
normal async function.

This patch adds and consolidates more Futures and Result type aliases
in the client.