chore: Only ButtplugServer should have to deal with message ids
The only place that set_id() should be called is in the message sorter
and ButtplugServer's parse_message() method. Nowhere else should have
to worry about it. Consolidate error types to make sure this happens,
and that conversion to/from ButtplugServerError is handled either in
parse_message() or in connectors.