bug: Youou wildcard prevents identifier from being found
The Youou devices have names like VX001_01234, where the numbers
after the _ are variable. To cope with this in the configuration
file, the protocol is matched using the wildcarded name VX001_*

When the default identifier lookup comes in though, it looks for
VX001_01234 which doesn't exist.

This change forces the identifier lookup to VX001_

An alternative but far larger change could be to have
DeviceConfigurationManager::find_configuration() report on the
identifier that was actually matched to see if wildcards were involved.
This would require changing the whole PartialEq implementation for
DeviceSpecifier, which given BLE names are not the only specifier
type that needs to be considered, didn't look like the right direction
whilst this issue only affects a small number of BLE protocols.