chore: Manually merging 3fb841bc from device-config repo
Turns out that buttplug-rs's device config is 5 commits
ahead of the device config repo and 1 behind. It might
be possible to do the subtree merge thing to pull these
in order, but the manual merge was trivial to bring over
the one missing change:

A few WeVibe devices were listed under the wevibe protocol
where they really needed to be listed under wevibe-8bit:
- Chorus
- Melt
- Wand

The Libo Feather also made its way into the config in that
change.