fix: Add event filtering and None handling to btleplug
Connect/disconnect events in BTLEPlug were not being filtered by
address. This means that any time a BTLEPlug device disconnected,
the event would be broadcast to all devices and they would all act
as if they had disconnected.

This also caused panics due to unwraps in the event handlers.

Fixes #197