chore: Replace lovense dongle unwrap with error/pass
For some reason, the serial dongle has timing issues where it can
try to write a device event before the listener is ready, meaning
we may not have a subscriber on the other end of our broadcast
channel yet. If this happens we *should* be able to ignore it, but
throw an error message anyways.

Fixes #280