fix: Splitting kiiroo_v21 for devices that need initailization
We know that the linear devices will accept the initialization without
any obious side-effects, whether needed or not.

The vibes on the otherhand will at best start up at 100% speed
or refuse all commands at worse.

I've left the vibration and linear logic in both protocols, so if
we've got a device in the wrong category, it's just a config change
to fix in the future.

Fixes #281