chore: Reduce log message level as temp Intiface Desktop fix
Intiface Desktop seems to get stalls in the engine executable when log
messages are posted from threads. This works fine when the CLI is used
in a terminal, but something about electron's execution model freaks
out here, even though this should just be sending to a channel. For
now, reduce the logging levels in threads on the lovense dongle.