feat: Adding support for Adult Festa Rocket+1D
Is there a typo in that vendor name? Nope... it really is festa...
  https://www.afesta.tv/gekisin/index2.html

As we know, Vorze products are always a little awkward due
to the device type being required in the control messages
(the Rocket+1D is 0x07), but this device was extra awkward
as it required writeWithResponse messages too.