fix: Using write with response on all Vorze devices
This fixes #353

Also added a test for the VorzePiston after correcting the
speed limit to 100. Sending >100 speed does seem to work,
but appears to run at the max speed of 100 regardless.
The piston logic doesn't handle incomplete moves (a movement
requested whilst the previous movement is still in progress)
so as a precaution, if told to move to the same location the
piston will do so at full speed rather than 0.