feat: Kizuna Smart support
This is a BT2 serial rotation controller that doesn't
have a reverse mode, so it's mapped as a scalar rotator.

The 2 outputs are not indivually controlled.

The device registers 2 RFCOMM ports, only the second seems
to allow connections.

Whilst I'm able to write to the device, I'm not getting any
response, but sirah on Discord confirmed that it worked for
them.

No tests, as I don't see an easy way to test serial devices.