feat: Allow for device reconnect
This change removes a device from the connecting list as soon as it's
done either connecting or having been released.

It's also possible for the device to removed prior to it being added to
device_map, which would trigger the expect() in the device removed
callback. I hit that in debugging, where things are much slower than
in reality, but I believe the race condition is still possible.