feat: Refactor device attribute handling in config manager
Move as much processing of device attributes back to the config
manager as possible. This includes changing how we load user
configurations in the backend, as well as separating long term (disk)
storage structures from internal (per-session) structures, while still
allowing short/long-term structures to convert back and forth easily.

In other words, store on the disk in a human usable form, and munge
everything to a somewhat weird tree when we load it.