fix: Additional fix-ups to the multi-protocol matcher
Seems like the device disconnect required an await to actually work.

A bigger problem was seen with my Hismith Cup: the service advert
come in so close to the name advert that the second event was always
skipped in handle_device_communication()/DeviceFound but the caller
(btplug_adapter_task::maybe_add_peripheral()) doesn't know the event
was dropped, so discards any repeated adverts without changes, which
means the Hismith is never picked up.

My solution to this is to loop in handle_device_communication()/DeviceFound
until either the device is no longer in the connecting state. If the
device connects, we exit.

Tested with:
* Lovense Lush
* Magic Motion Curve
* Hismith Cup
* Satisfyer Hot Lover