feat: Support Satisfyer devices with more than 2 vibes
We've been sending packets of 8 bytes to all Satisfyer devices
so far, but it turns out that we should be sending 4 bytes per
feature. Right now the features are vibrators, oscilators or
single direction rotators (all map to VibrateCmd).

This change uses the feature count to initialise the length of
the data packet array, and dynamically loads the values into the
array before sending.