chore: Make server/devicemgr use builders, immutable after construction
There was a lot of assumed mutability during a server lifetime that was
never really possible to actually use, and was making the system far
more complicated than it needed to be. Move server and child mgrs
to use builders, and once created, make immutable. This saves a lot
of code, internal mutability worries, etc...