chore: Update prost and prost-build to 0.11
This fixes cross-compilation from Linux to Windows. prost-build 0.10 has
a bug that prevents it from working.

However, prost-build 0.11 comes with a new build requirement where the
environment needs to provide protoc. Previously, prost-build would try
to build protoc from source if it was not available. buttplug, in turn,
inherits this requirement as such.