feat: Passing ProtocolDeviceAttributes to ProtocolInitializer::initialize()
This allows protocol initialisation to be feature aware.

The included use-case is the Satisfyer protocol which sends the last command
sent as a keep-alive, starting in the initialiser. The commands are variable
length based on the capabilities of the device, so we need to know if the
device has 1 to 5 features to properly setup the keep-alive.