- Implemented ProtocolHandler handle_scalar_cmd method for LovenseConnectService.
I wrote a lot of notes in this trait method. They are some things to think about based of my testing results.
The TLDR is that the spec isnt fully supported by LovenseConnect (based off my testing).
So I have written in two other "hacks" that can be uncommented and used to make Constriction support for LovenseConnect not crappy.
Hack 1: Using the unused steps (0,4,5) as subcommands to the LovenseConnect app.
Hack 2: Use the LovenseConnect AirAuto command to create an "on" toggle where the toy
will infinitely constrict and stay the most constricted it can be until a level 5 step is sent.
Then an AirOut command will be sent which wil toggle "off" the constriction.

~ SutekhVRC