chore: Remove ButtplugRemoteServer
Move module up to intiface-engine, as that's the only place we use it.
Anyone else who will use this probably has their own impl, so this will
provide us extra freedom to randomly change whatever we want while not
having to rev the library.

Introduces ButtplugTestServer for tests, which is mostly the same
as remote server. Could probably be paired down and cleaned up but eh.