feat: Adding support for Fredorch Rotary devices
Note: these devices are almost as bad as fire and forget RF. We're tracking our
assumed state of the device and updating the speed based on what amount to hitting
the speed up/down buttons. Id the device was previoulsy connected, it may already
ben in the "powered on" state, in which case out initialiser will change the state
to off (the device will beep twice) but we won't know and it'll be up to the user
reset the device (or restart the connection somehow).