feat: Adding support for Lovense Vulse and Lapis
The Lapis is doing something odd where it's not recieving notifications
on Windows: meaning DeviceType isn't usable. Instead the device name is
used as a fallback (using a regex)

Lapis "works" with Lovense Connect, except that I can't get the Vibrate3
command to work, so it;s just giver/receiver vibes when connected this way.