chore: Implement proper message versioning
This is a *huge* commit that implements proper message versioning
instead of assuming that ButtplugClientMessage/ButtplugServerMessage
will cover our needs for all message versions. This is required due to
message spec v4 changing messages in a way that will require complex
conversion. However, it should make future updates of the message
spec easier to implement, since all message backward compat work is now
implemented in the same submodule, and all library logic is considered
to work on (mostly) the same spec version.

Note that the client spec version for this commit is still v3 as I am
planning on releasing the server with v4 capabilities to test the code
before publishing the client and spec.