chore: Divide into truly core/client/server modules
We've had a ton of client-only or server-only implementations in core
since we moved to rust, with no good reasoning for it. Core should ONLY
include things that both client and server need, and anything client or
server specific should be moved to those modules.

This means that all old messages not in the current v4 spec are being
moved into the server module, and connector/serializers for specific
client/server needs are being moved into their respective modules.