fix: Port range fix (#961) to v4 message conversion
This conversion originally happened in the actuator command manager, but
was moved to message conversion. A fix happened to make sure that we
heed the step limit versus the range in the v3 line ACM, and needed to
be ported into this conversion to make sure the test still passes.